backend/requirements.txt
APIFlask==2.2.0
colorlog==6.8.2
elevenlabs==1.7.0
Flask==3.0.3
Flask_Cors==4.0.1
Flask_SocketIO==5.3.6
flask_sqlalchemy==3.1.1
marshmallow_dataclass==8.7.0
openai==1.41.1
git+https://github.com/ground-creative/openvoice-api-client-python.git
pydub==0.25.1
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.32.3
SQLAlchemy==2.0.32
twilio==9.2.3
Werkzeug==3.0.3
mysql-connector-python==9.0.0
